Can't Sync Visual Studio After Merge (Again)
