+++ Game1/Content/Content.mgcb
@@ -464,3 +464,15 @@
 /processorParam:TextureFormat=Color
 /build:BGFinalFantasy.jpg
 
+#begin Cabin1.png
+/importer:TextureImporter
+/processor:TextureProcessor
+/processorParam:ColorKeyColor=255,0,255,255
+/processorParam:ColorKeyEnabled=True
+/processorParam:GenerateMipmaps=False
+/processorParam:PremultiplyAlpha=True
+/processorParam:ResizeToPowerOfTwo=False
+/processorParam:MakeSquare=False
+/processorParam:TextureFormat=Color
+/build:Cabin1.png
+
