+++ Game1/Land.cs
@@ -34,12 +34,13 @@ namespace Game1
         //public int Y;
         //public int land { get; set; }
         //public int mod { get; set; }
+        //public float? rotate = null;
         public int frame = 5;
         public int biome = 0;
         public int land = 0;
-        public float? rotate = null;
         public bool IsActive = false;
         public bool IsOccupied = false;
+        public DateTime LastUpdate = new DateTime();
 
         public Land()//int x, int y, int m)
         {
